Update sqlHelper to avoid returning sql scripts as string.
